--- Minning/ViewControllers/Main/Report/Views/FilterButton.swift
@@ -1,9 +1,0 @@
-//
-//  FilterButton.swift
-//  Minning
-//
-//  Created by denny on 2021/10/20.
-//  Copyright © 2021 Minning. All rights reserved.
-//
-
-import Foundation
