+++ Minning/ViewControllers/Main/Report/Views/WeeklyPercentView.swift
@@ -1,0 +1,32 @@
+//
+//  WeeklyPercentView.swift
+//  Minning
+//
+//  Created by denny on 2021/10/03.
+//  Copyright © 2021 Minning. All rights reserved.
+//
+
+import CommonSystem
+import DesignSystem
+import Foundation
+import SharedAssets
+import SnapKit
+
+final class WeeklyPercentView: UIView {
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setupView()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    private func setupView() {
+        // SAMPLE
+        backgroundColor = .systemOrange
+        self.snp.makeConstraints { make in
+            make.height.equalTo(200)
+        }
+    }
+}
