--- SharedAssets/SharedAssets/Colors/ColorSet.xcassets/gray787C84.colorset/Contents.json
@@ -1,20 +1,0 @@
-{
-  "colors" : [
-    {
-      "color" : {
-        "color-space" : "srgb",
-        "components" : {
-          "alpha" : "1.000",
-          "blue" : "0x84",
-          "green" : "0x7C",
-          "red" : "0x78"
-        }
-      },
-      "idiom" : "universal"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
