+++ SharedAssets/SharedAssets/Colors/ColorSet.xcassets/minningLightGray100.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "srgb",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0xF8",
-          "green" : "0xF7",
-          "red" : "0xF6"
+          "blue" : "0.976",
+          "green" : "0.969",
+          "red" : "0.965"
         }
       },
       "idiom" : "universal"
